requirements.txt
black==24.8.0
Django==5.1.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
twilio==9.3.1